requirements.txt
Django==4.2.16
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.2.0
ultralytics>=8.2.0
opencv-python==4.8.1.78
easyocr==1.7.1
numpy==1.24.3
matplotlib==3.7.5